move pmtiles to the new location
