+++ martin-core/Cargo.toml
@@ -30,10 +30,11 @@ sprites = [
 ]
 styles = ["dep:walkdir", "tokio/fs", "dep:dashmap"]
 mbtiles = ["dep:mbtiles"]
-pmtiles = ["dep:pmtiles"]
+pmtiles = ["dep:pmtiles", "dep:aws-config", "dep:url"]
 
 [dependencies]
 async-trait.workspace = true
+aws-config = { workspace = true, optional = true }
 bit-set = { workspace = true, optional = true }
 dashmap = { workspace = true, optional = true }
 futures = { workspace = true, optional = true }
@@ -54,6 +55,7 @@ thiserror.workspace = true
 tiff = { workspace = true, optional = true }
 tilejson.workspace = true
 tokio = { workspace = true, optional = true }
+url = { workspace = true, optional = true }
 walkdir = { workspace = true, optional = true }
 
 [dev-dependencies]
